moving checkMessageDecryption to its own file in utils
