+++ src/tests/utils/msg/sendTribeMessageAndCheckDecryption.ts
@@ -1,30 +1,10 @@
 import { Assertions } from 'ava'
 import { NodeConfig } from '../../types'
-import * as rsa from '../../../crypto/rsa'
-import { getCheckNewMsgs } from '../get'
 import { Message, Chat } from '../../types'
-import { sendTribeMessage } from '../msg'
+import { sendTribeMessage, checkMessageDecryption } from '../msg'
 
 interface SendMessageOptions {
   amount: number
-}
-
-export async function checkDecryption(
-  t: Assertions,
-  node: NodeConfig,
-  msgUuid,
-  text
-) {
-  // //wait for message to process
-  const lastMessage = await getCheckNewMsgs(t, node, msgUuid)
-  t.truthy(lastMessage, 'await message post')
-
-  //decrypt the last message sent to node using node private key and lastMessage content
-  const decrypt = rsa.decrypt(node.privkey, lastMessage.message_content)
-
-  //the decrypted message should equal the random string input before encryption
-  t.true(decrypt === text, 'decrypted text should equal pre-encryption text')
-  return true
 }
 
 // send a message
@@ -43,7 +23,7 @@ export async function sendTribeMessageAndCheckDecryption(
 
   const msgUuid = msg.uuid
 
-  await checkDecryption(t, node2, msgUuid, text)
+  await checkMessageDecryption(t, node2, msgUuid, text)
 
   return msg
 }
